+++ tests/config/Models/Enums/StateEnum.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Backpack\CRUD\Tests81\Unit\Models\Enums;
+namespace Backpack\CRUD\Tests\config\Models\Enums;
 
 enum StateEnum
 {
